+++ runtime/c/api/packet.h
@@ -11,8 +11,8 @@
 #include <stdbool.h>
 
 typedef struct {
-    int model_id;
-    int record_id;
+    uint8_t model_id;
+    uint16_t record_id;
 
     bool is_ack;
     bool is_save_done;
@@ -23,9 +23,9 @@ typedef struct {
     size_t packet_length;
 } RavelPacket;
 
-void ravel_packet_init_empty (RavelPacket *self, size_t record_size, int model_id, int record_id);
-void ravel_packet_init_ack (RavelPacket *self, int model_id, int record_id);
-void ravel_packet_init_save_done (RavelPacket *self, int model_id, int record_id);
+void ravel_packet_init_empty (RavelPacket *self, size_t record_size, uint8_t model_id, uint16_t record_id);
+void ravel_packet_init_ack (RavelPacket *self, uint8_t model_id, uint16_t record_id);
+void ravel_packet_init_save_done (RavelPacket *self, uint8_t model_id, uint16_t record_id);
 
 void ravel_packet_init_from_record (RavelPacket *self, uint8_t *data, size_t length);
 void ravel_packet_init_from_network (RavelPacket *self, uint8_t *data, size_t length);
